# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TreePage API
  version: 0.1.0
  description: Documentation platform REST API
servers:
  - url: http://localhost:8081
    description: Auth service (dev)
  - url: http://localhost:8082
    description: Server API (dev)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        is_public:
          type: boolean
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        space_id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        path:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        author_name:
          type: string
        updated_at:
          type: string
          format: date-time
    SearchResult:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        snippet:
          type: string
        rank:
          type: number
paths:
...
  /api/spaces:
    get:
      tags: [Spaces]
      summary: List accessible spaces
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Space"
    post:
      tags: [Spaces]
      summary: Create space
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name]
              properties:
                slug:
                  type: string
                name:
                  type: string
                description:
                  type: string
                is_public:
                  type: boolean
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
  /api/spaces/{slug}:
    get:
      tags: [Spaces]
      summary: Get space by slug
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Space"
  /api/spaces/{slug}/documents:
    get:
      tags: [Documents]
      summary: List documents in space
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Document"
  /api/spaces/{slug}/documents/{docSlug}:
    get:
      tags: [Documents]
      summary: Get document
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: docSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /api/search:
    get:
      tags: [Search]
      summary: Full-text search
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: space_id
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tags
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchResult"
                  total:
                    type: integer
...
  /api/rag/ask:
    post:
      tags: [RAG]
      summary: Ask documentation (RAG)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 10
      responses:
        "200":
          description: Answer with sources and citations
  /api/rag/feedback:
    post:
      tags: [RAG]
      summary: Submit RAG answer feedback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question, helpful]
              properties:
                question:
                  type: string
                answer:
                  type: string
                helpful:
                  type: boolean
                confidence:
                  type: number
      responses:
        "200":
          description: Feedback recorded
  /api/documents/{id}/sync-diff:
    get:
      tags: [Documents]
      summary: Diff local edits vs last Git sync snapshot
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Line diff between Git snapshot and local content
...